move metadata utility to deprecated namespace
