+++ db/tests/queries/conftest.py
@@ -2,7 +2,7 @@ import pytest
 from sqlalchemy import inspect
 from db.deprecated.columns import get_column_attnum_from_name as get_attnum
 from db.queries.base import DBQuery, InitialColumn, JoinParameter
-from db.metadata import get_empty_metadata
+from db.deprecated.metadata import get_empty_metadata
 
 
 @pytest.fixture
