+++ db/tests/queries/test_base.py
@@ -2,7 +2,7 @@ from sqlalchemy import inspect
 from db.queries.base import DBQuery, InitialColumn, JoinParameter
 from db.deprecated.columns import get_column_attnum_from_name as get_attnum
 from db.transforms import base as tbase
-from db.metadata import get_empty_metadata
+from db.deprecated.metadata import get_empty_metadata
 
 
 def _extract_col_properties_dict(col):
