+++ db/tests/queries/test_initial_relation.py
@@ -6,7 +6,7 @@
 from sqlalchemy import inspect
 from db.deprecated.columns import get_column_attnum_from_name as get_attnum
 from db.queries.base import DBQuery, InitialColumn, JoinParameter
-from db.metadata import get_empty_metadata
+from db.deprecated.metadata import get_empty_metadata
 
 
 def _get_oid_from_table(name, schema, engine):
