+++ db/tests/transforms/test_finish_specifying_summarization.py
@@ -4,7 +4,7 @@ import types
 import functools
 
 from db.queries.base import DBQuery, InitialColumn, JoinParameter
-from db.metadata import get_empty_metadata
+from db.deprecated.metadata import get_empty_metadata
 from db.transforms.base import Summarize, SelectSubsetOfColumns, Limit
 from db.deprecated.columns import get_column_attnum_from_name
 from db.transforms.operations.finish_specifying import finish_specifying_summarize_transform
